# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Medioa REST API (with gin-gonic).",
        "title": "Medioa API",
        "contact": {
            "name": "Vũ Kỳ",
            "url": "github.com/vukyn",
            "email": "vukynpro@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/share/download/{file_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Download media file (images, videos, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Share"
                ],
                "summary": "Download media (public/private)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.DownloadResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "medioa_internal_storage_models.DownloadResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}